# api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/submission/submit": {
      "post": {
        "operationId": "submitHoldings",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmissionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "submission"
        ]
      }
    },
    "/api/submission/status/{address}": {
      "get": {
        "operationId": "getSubmissionStatus",
        "parameters": [
          {
            "name": "address",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmissionStatusDto"
                }
              }
            }
          }
        },
        "tags": [
          "submission"
        ]
      }
    },
    "/api/submission/submit-csv": {
      "post": {
        "operationId": "submitCustomersHoldingsCsv",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "submission"
        ]
      }
    },
    "/api/customer/verify": {
      "post": {
        "operationId": "verifyHoldings",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerificationDto"
                }
              }
            }
          }
        },
        "tags": [
          "customer"
        ]
      }
    },
    "/api/crypto/balance/{address}": {
      "get": {
        "operationId": "getBalance",
        "parameters": [
          {
            "name": "address",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "crypto"
        ]
      }
    },
    "/api/system/config": {
      "get": {
        "operationId": "getSystemConfig",
        "parameters": [],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemConfig"
                }
              }
            }
          }
        },
        "tags": [
          "system"
        ]
      }
    },
    "/api/system": {
      "get": {
        "operationId": "systemTest",
        "parameters": [],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemStatus"
                }
              }
            }
          }
        },
        "tags": [
          "system"
        ]
      }
    },
    "/api/test/reset": {
      "get": {
        "operationId": "resetDb",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "test"
        ]
      }
    },
    "/api/test/send-test-email": {
      "post": {
        "operationId": "sendTestEmail",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendTestEmailDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "test"
        ]
      }
    },
    "/api/test/send-test-verification-email": {
      "post": {
        "operationId": "sendTestVerificationEmail",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendTestEmailDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "test"
        ]
      }
    },
    "/api/test/send-funds": {
      "post": {
        "operationId": "sendFunds",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendFundsDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "test"
        ]
      }
    },
    "/api/exchange": {
      "get": {
        "operationId": "getAllExchanges",
        "parameters": [],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ExchangeDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "exchange"
        ]
      }
    }
  },
  "info": {
    "title": "",
    "description": "",
    "version": "1.0.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "schemas": {
      "CustomerHoldingDto": {
        "type": "object",
        "properties": {
          "hashedEmail": {
            "type": "string"
          },
          "amount": {
            "type": "number"
          }
        },
        "required": [
          "hashedEmail",
          "amount"
        ]
      },
      "SubmissionDto": {
        "type": "object",
        "properties": {
          "exchangeName": {
            "type": "string"
          },
          "customerHoldings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CustomerHoldingDto"
            }
          }
        },
        "required": [
          "exchangeName",
          "customerHoldings"
        ]
      },
      "SubmissionStatus": {
        "type": "string",
        "enum": [
          "unused",
          "waiting-for-payment",
          "complete"
        ]
      },
      "SubmissionStatusDto": {
        "type": "object",
        "properties": {
          "paymentAddress": {
            "type": "string"
          },
          "paymentAmount": {
            "type": "number"
          },
          "submissionStatus": {
            "$ref": "#/components/schemas/SubmissionStatus"
          }
        },
        "required": [
          "paymentAddress",
          "paymentAmount",
          "submissionStatus"
        ]
      },
      "EmailDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          }
        },
        "required": [
          "email"
        ]
      },
      "VerificationResult": {
        "type": "string",
        "enum": [
          "email-sent",
          "failed-to-send-email",
          "cant-find-holdings-for-email",
          "cant-find-verified-holdings-for-email"
        ]
      },
      "VerificationDto": {
        "type": "object",
        "properties": {
          "verificationResult": {
            "$ref": "#/components/schemas/VerificationResult"
          }
        },
        "required": [
          "verificationResult"
        ]
      },
      "SystemConfig": {
        "type": "object",
        "properties": {
          "docsUrl": {
            "type": "string"
          }
        },
        "required": [
          "docsUrl"
        ]
      },
      "SystemStatus": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          }
        },
        "required": [
          "status"
        ]
      },
      "SendTestEmailDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          }
        },
        "required": [
          "email"
        ]
      },
      "SendFundsDto": {
        "type": "object",
        "properties": {
          "fromAddress": {
            "type": "string"
          },
          "toAddress": {
            "type": "string"
          },
          "amount": {
            "type": "number"
          }
        },
        "required": [
          "fromAddress",
          "toAddress",
          "amount"
        ]
      },
      "ExchangeDto": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "exchangeName": {
            "type": "string"
          }
        },
        "required": [
          "_id",
          "exchangeName"
        ]
      }
    }
  }
}